requirements.txt
cfgv==3.4.0
distlib==0.3.7
filelock==3.12.3
identify==2.5.27
nodeenv==1.8.0
platformdirs==3.10.0
pre-commit==3.4.0
PyYAML==6.0.1
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
six==1.16.0
virtualenv==20.24.5
